Migrate task pages to TS
