--- client/src/components/TaskContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const TaskContext = createContext({});
