+++ client/src/components/TaskContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { TaskContextType } from "../types";
+
+export const TaskContext = createContext<TaskContextType | undefined>(
+  undefined
+);
